bio/models.py
from django.db import models
from ckeditor.fields import RichTextField

# Create your models here.
class Category(models.Model):
    name = models.CharField(max_length=250, unique=True)
    slug = models.SlugField(max_length=250, unique=True)

    def __str__(self):
        return self.name
    

class News(models.Model):
    category = models.ForeignKey(Category, on_delete=models.PROTECT)
    created_at = models.DateTimeField(auto_now_add=True)
    updated_at = models.DateTimeField(auto_now=True)
    title = models.CharField(max_length=250)        
    slug = models.SlugField(max_length=250, unique=True)
    summary = RichTextField()
    description = RichTextField()
    is_published = models.BooleanField(default=True)
    photo = models.ImageField(upload_to='photos/%y/%m/%d/', blank=True)
    views = models.IntegerField(default=0)

    def __str__(self):
        return self.title


class Page(models.Model):
    title = models.CharField(max_length=255, unique=True)
    slug = models.SlugField(max_length=255, unique=True)
    image = models.ImageField(upload_to = 'page/%y/%m/%d/', blank=True)
    content = RichTextField()   

    def __str__(self):
        return self.title     

